stdio: bypass uartstdio on native boards where libc's facilities are used
on native, there is no implementation of uartstdio, and rather than that
providing a backend to libc's stdio system, libc is used unmodified.

as the object in stdio tries to access riot's io facilities directly
(which might or might not be a good idea), it needs to resort to direct
c printing (or stdlib, but i'd prefer to avoid that) on native.