threads: Change API to be safe against moving of the Thread struct
The Thread struct can now freely be moved -- as it takes references to
both the stack and the closure. This was chanegd because it turned out
that the Pin methods can not keep the returned object from moving on the
stack, so the two-phase approach of preparation and starting to have a
stable pointable-to self was pointless (no pun intended).

The new API is not without issues of its own (ergonomics suffered), but
at least doesn't depend on impossible pinning any more.

The problem of a therad not being allowed to be forgotten is unaffected
by this and still unsolved.