RIOT version compatibility improvements
This introduces checks to make the Rust wrappers usable even before the
recently adapted-to change in the thread_{status,state}_t name, and
reflecs the more recent thread priority type while staying compatible
with older versions.

Where required, those compatibility adaptions use newly introduced cfg
options for certain RIOT version thresholds (comparison is done in the
build scripts, as cfg values would be string valued and are unsorted;
flags are introduced whenever there is a new fix to be considered, and
removed when compatibility with older RIOT versions is dropped).