shell: Change arguments type
This allows removing the arbitrary limit on argc. This is a breaking
change, but all the regular operations to be expected on the previously
passed slice can just as well be done on this new type.

The previous implementation was based on the erroneous impression that
in order to hand out the &str, an (address, length) that backs the
str would need to be stored internally (and that can't be done for
arbitrarily many arguments) -- when in fact, the length is part of the
&str type and is passed out as a value together with the reference.