Shell: Rewrite
This changes the API completely, but the migration is rather
straightforward as shown with the examples.

Not only does this get rid of the limited size command count in favor of
a type grown by the builder, it also replaces the string comparison that
was necessary with the old one (that had a shared callback for all
commands and went over argv[0] again) with trampolines. (They still have
to go through a round dispatch where the right self has to be picked out
from the list of selves in the commands root, but that's only a matter
of simple comparisons).