cstr_core: Make nightly feature optional, deprecating direct use
This deprecation is worth pointing out in the change log as it's not
automatically visible: The use of nightly features of cstr_core is
deprecated, but only if the new cstr_nightly feature is not enabled.
(So users, test your builds with the new feature off. If all builds,
you're good, and you may want to stay on manual-defaults until the next
breaking release. If not, go back to regular defaults and enable the
cstr_nightly feature.