doc: Ensure buildability
The main example is still not buildable, but spews more compact errors
now. Generally, examples building needs -Z doctest-xcompile, which (at
least as used in riot-doc-helpers) still appears to attempt to run the
test. Thus, testing all this is not automatable yet, but was manually
checked to produce errors as late as practical (usually at attempted
runtime).