GNRC: Change return types, and ports to NonZero
Several GNRC components used Option<T> or Result<T, ()> instead of a
more descriptive error type (NotEnoughSpace) that has been introduced
for new functions but not changed previously to preserve the API.

In the course of more precise error handling, UDP ports were changed to
be NonZeroU16 (as sending from or to 0 is not accepted by GNRC).
Consequently (as the stack does not check incoming packets), the
`from_incoming` method became fallible -- which is generally a good
thing, because not all incoming packets can be responded to in general.
(Think packets to multicast -- these need more elaborate handling and a
decision where to respond from).