CI: Run everything in parallel
The largest amount of time is the machine checkout -- 2 minutes. The 7
seconds of `cargo update` are dwarved by that, and running in parallel
means we don't have to eat those 2 minutes twice.

(Also, I couldn't figure out how to carry the output state of the
`needs`'d preparations over to the actual working steps).