gpio: Provide direct methods in addition to HAL
This adds a .set_state() call (which would otherwise not be available as
it's a provided method of the traits), which also benefits HAL users
because the direct implementation can do without the branching of the
provided method.

Test pinouts are also provided for microbit-v2 because that's where it
was just tested.