panic: Crash sooner with panic_handler_crash
With the feature panic_handler_crash, a panic would previously have
tried to print details (depending on other features) before calling the
RIOT panic. This feature's purpose is to shut down everything ASAP, so
instead it now also bypasses these prints, allowing the compiler to also
remove the check for whether an interrupt is active.