Change `!` in associated error types to Infallible
The `Never` helper still exists, because it is necessary to implement
traits such as Termination for `!`, and to express that an input closure
must not terminate, but the public error types are changed.

While this is a breaking change that was not preceded by a deprecation,
it is unlikely people match on the precise type (and all error
conversions should be implemented or implementable on both types).