gpio: Refactor use of RIOT's gpio_read()
C callers of gpio_read() do not care whether gpio_read() returns an
pre C99 style `bool` (which is an `int` with every non-zero number
being a valid representation of `true`) or C99 style `bool`, as long
as the stay within spec (treating representation of `true` as `true`
with no regard to the encoding).

This refactors the rust use to match the C behavior. This would allow
updating `gpio_read()` to return C99 style `bool` as type.